optimized Model object
