+++ pyNTM/model.py
@@ -14,6 +14,7 @@ from pprint import pprint
 
 import itertools
 import networkx as nx
+import random
 
 from .circuit import Circuit
 from .interface import Interface
@@ -246,63 +247,155 @@ class Model(MasterModel):
 
         self.validate_model()
 
-    # def _route_demands(self, model):
-    #     """
-    #     Routes demands in input 'model'
-    #
-    #     :param model: input 'model' parameter object (may be different from self)
-    #     :return: model with routed demands
-    #     """
-    #
-    #     G = self._make_weighted_network_graph_mdg(include_failed_circuits=False)
-    #
-    #     for demand in model.demand_objects:
-    #         src = demand.source_node_object.name
-    #         dest = demand.dest_node_object.name
-    #
-    #         # Shortest path in networkx multidigraph
-    #         try:
-    #             nx_sp = list(nx.all_shortest_paths(G, src, dest, weight='cost'))
-    #         except nx.exception.NetworkXNoPath:
-    #             # There is no path, demand.path = 'Unrouted'
-    #             demand.path = 'Unrouted'
-    #             continue
-    #
-    #         # Find demand's load to interfaces it touches
-    #         num_paths = len(nx_sp)
-    #         demand_load_per_path = demand.traffic / num_paths
-    #
-    #         # all_paths is list of paths from source to destination; these paths
-    #         # may include paths that have multiple links between nodes
-    #         all_paths = []
-    #
-    #         for path in nx_sp:
-    #             current_hop = path[0]
-    #             this_path = []
-    #             for next_hop in path[1:]:
-    #                 this_hop = []
-    #                 values_source_hop = G[current_hop][next_hop].values()
-    #                 min_weight = min(d['cost'] for d in values_source_hop)
-    #                 ecmp_links = [interface_index for interface_index, interface_item in
-    #                               G[current_hop][next_hop].items() if
-    #                               interface_item['cost'] == min_weight]
-    #                 num_ecmp_links = len(ecmp_links)
-    #
-    #                 # Add Interface(s) to this_hop list and add traffic to Interfaces
-    #                 for link_index in ecmp_links:
-    #                     this_hop.append(G[current_hop][next_hop][link_index]['interface'])
-    #                 this_path.append(this_hop)
-    #                 current_hop = next_hop
-    #             all_paths.append(this_path)
-    #
-    #         path_list = self._normalize_multidigraph_paths(all_paths)
-    #         demand.path = path_list
-    #
-    #     self._update_interface_utilization()
-    #
-    #     return self
+    def _route_demands(self, model):
+        """
+        Routes demands in input 'model'
 
-    def _make_weighted_network_graph_mdg(self, include_failed_circuits=True, needed_bw=0, rsvp_required=False):
+        :param model: input 'model' parameter object (may be different from self)
+        :return: model with routed demands
+        """
+
+        G = self._make_weighted_network_graph(include_failed_circuits=False)
+
+        for demand in model.demand_objects:
+            demand.path = []
+
+            # Find all LSPs that can carry the demand:
+            for lsp in (lsp for lsp in model.rsvp_lsp_objects):
+                if (lsp.source_node_object == demand.source_node_object and
+                        lsp.dest_node_object == demand.dest_node_object and
+                        'Unrouted' not in lsp.path):
+                    demand.path.append(lsp)
+
+            if demand.path == []:
+                src = demand.source_node_object.name
+                dest = demand.dest_node_object.name
+
+                # Shortest path in networkx multidigraph
+                try:
+                    nx_sp = list(nx.all_shortest_paths(G, src, dest, weight='cost'))
+                except nx.exception.NetworkXNoPath:
+                    # There is no path, demand.path = 'Unrouted'
+                    demand.path = 'Unrouted'
+                    continue
+
+                # all_paths is list of paths from source to destination
+                all_paths = self.convert_graph_path_to_model_path(nx_sp)
+
+                demand.path = all_paths
+
+        self._update_interface_utilization()
+
+        return self
+
+    def _determine_lsp_state_info(self, lsps, traff_on_each_group_lsp):
+        """
+        Determine LSP's specific path and reserved bandwidth; also consume
+        reserved bandwidth on transited Interfaces
+
+        :param lsps: List of parallel LSPs (LSPs with common source/dest nodes)
+        :param traff_on_each_group_lsp: How much traffic each LSP should attempt
+        to carry
+        :return: None; determines path and reserved bandwidth for each LSP in lsps
+        and also consumes reservable bandwidth on each Interface each LSP transits
+        """
+
+        for lsp in lsps:
+            # Check to see if configured_setup_bandwidth is set; if so,
+            # set reserved_bandwidth and setup_bandwidth equal to
+            # configured_setup_bandwidth value
+            if lsp.configured_setup_bandwidth is None:
+                lsp.reserved_bandwidth = traff_on_each_group_lsp
+                lsp.setup_bandwidth = traff_on_each_group_lsp
+            else:
+                lsp.reserved_bandwidth = lsp.configured_setup_bandwidth
+                lsp.setup_bandwidth = lsp.configured_setup_bandwidth
+
+            G = self._make_weighted_network_graph(include_failed_circuits=False, rsvp_required=True,
+                                                  needed_bw=lsp.setup_bandwidth)
+
+            lsp.path = {}
+
+            # Get shortest paths in networkx multidigraph; for DiGraph, only edge names will be
+            # in returned paths (the rest of the edge data will not)
+            try:
+                nx_sp = list(nx.all_shortest_paths(G, lsp.source_node_object.name, lsp.dest_node_object.name,
+                                                   weight='cost'))
+            except nx.exception.NetworkXNoPath:
+                # There is no path; path = 'Unrouted'
+                lsp.path = 'Unrouted'
+                lsp.reserved_bandwidth = 'Unrouted'
+                continue
+
+            # Convert node hop by hop paths from G into Interface-based paths
+            all_paths = self.convert_graph_path_to_model_path(nx_sp)
+
+            # # all_paths may have hops between nodes that can take different Interfaces;
+            # # normalize those hops that could transit any of multiple Interfaces into
+            # # distinct, unique possible paths
+            # candidate_path_info = self._normalize_multigraph_paths(all_paths)
+
+            # Candidate paths with enough reservable bandwidth
+            candidate_path_info_w_reservable_bw = []
+
+            # Determine which candidate paths have enough reservable bandwidth
+            for path in all_paths:
+                if min([interface.reservable_bandwidth for interface in path]) >= lsp.setup_bandwidth:
+                    candidate_path_info_w_reservable_bw.append(path)
+
+            # If multiple lowest_metric_paths, find those with fewest hops
+            if len(candidate_path_info_w_reservable_bw) == 0:
+                lsp.path = 'Unrouted'
+                lsp.reserved_bandwidth = 'Unrouted'
+                continue
+
+            elif len(candidate_path_info_w_reservable_bw) > 1:
+                fewest_hops = min([len(path) for path in candidate_path_info_w_reservable_bw])
+                lowest_hop_count_paths = [path for path in candidate_path_info_w_reservable_bw
+                                          if len(path) == fewest_hops]
+                if len(lowest_hop_count_paths) > 1:
+                    new_path = random.choice(lowest_hop_count_paths)
+                else:
+                    new_path = lowest_hop_count_paths[0]
+            else:
+                new_path = candidate_path_info_w_reservable_bw[0]
+
+            # Change LSP path into more verbose form and set LSP's path
+            self._add_lsp_path_data(lsp, new_path)
+
+            for interface in [interface for interface in lsp.path['interfaces'] if lsp.path != 'Unrouted']:
+                interface.reserved_bandwidth += lsp.reserved_bandwidth
+
+    def convert_graph_path_to_model_path(self, nx_sp):
+        """
+        Converts list of Node names on path to list of Interfaces in Model on path.  This
+        can only be used on MultiGraph paths because MultiGraphs only support a single edge
+        (Circuit) between Node objects.
+
+        :param nx_sp: List of Node name hops on path
+        Example: [['A', 'B', 'D'], ['A', 'B', 'G', 'D']]
+        :return: List of Interface objects on path
+        Example: for the ['A', 'B', 'D'] path above, the converted path would be
+        [Interface(name = 'A-to-B', cost = 20, capacity = 125.0, node_object = Node('A'),
+            remote_node_object = Node('B'), circuit_id = None),
+         Interface(name = 'B-to-D', cost = 20,
+            capacity = 125.0, node_object = Node('B'), remote_node_object = Node('D'), circuit_id = None)]
+
+        """
+        all_paths = []
+        for nx_graph_path in nx_sp:
+            model_path = []
+            for hop in nx_graph_path:
+                current_hop_index = nx_graph_path.index(hop)
+                next_hop_index = current_hop_index + 1
+                if next_hop_index < len(nx_graph_path):
+                    next_hop = nx_graph_path[next_hop_index]
+                    interface = self.get_interface_object_from_nodes(hop, next_hop)
+                    model_path.append(interface)
+            all_paths.append(model_path)
+        return all_paths
+
+    def _make_weighted_network_graph(self, include_failed_circuits=True, needed_bw=0, rsvp_required=False):
         """
         Returns a networkx weighted networkx multidigraph object from
         the input Model object
@@ -316,7 +409,7 @@ class Model(MasterModel):
         rsvp_required parameters
         """
 
-        G = nx.MultiDiGraph()
+        G = nx.DiGraph()
 
         # Get all the edges that meet 'failed' and 'reservable_bw' criteria
         if include_failed_circuits is False:
@@ -348,9 +441,9 @@ class Model(MasterModel):
 
         return G
 
-    def _normalize_multidigraph_paths(self, path_info):  # TODO - static?
+    def _normalize_multigraph_paths(self, path_info):  # TODO - static?
         """
-        Takes the multidigraph_path_info and normalizes it to create all the
+        Takes the multigraph_path_info and normalizes it to create all the
         path combos that only have one link between each node.
 
         :param path_info: List of of interface hops from a source
@@ -361,16 +454,11 @@ class Model(MasterModel):
         path_info example from source node 'B' to destination node 'D':
         [
             [[Interface(name = 'B-to-D', cost = 20, capacity = 125, node_object = Node('B'),
-                    remote_node_object = Node('D'), circuit_id = '3')]], # there is 1 interface from B to D and a
-                    complete path
+                    remote_node_object = Node('D'), circuit_id = '3')]],
             [[Interface(name = 'B-to-G_3', cost = 10, capacity = 100, node_object = Node('B'),
-                    remote_node_object = Node('G'), circuit_id = '28'),
-              Interface(name = 'B-to-G', cost = 10, capacity = 100, node_object = Node('B'),
-                    remote_node_object = Node('G'), circuit_id = '8'),
-              Interface(name = 'B-to-G_2', cost = 10, capacity = 100, node_object = Node('B'),
-                    remote_node_object = Node('G'), circuit_id = '18')], # there are 3 interfaces from B to G
-            [Interface(name = 'G-to-D', cost = 10, capacity = 100, node_object = Node('G'),
-                    remote_node_object = Node('D'), circuit_id = '9')]] # there is 1 int from G to D; end of 2nd path
+                    remote_node_object = Node('G'), circuit_id = '28')],
+            [[Interface(name = 'G-to-D', cost = 10, capacity = 100, node_object = Node('G'),
+                    remote_node_object = Node('D'), circuit_id = '9')]]
         ]
 
 
@@ -417,7 +505,7 @@ class Model(MasterModel):
                  comprised of two Interface objects
         """
 
-        G = self._make_weighted_network_graph_mdg(include_failed_circuits=include_failed_circuits)
+        G = self._make_weighted_network_graph(include_failed_circuits=include_failed_circuits)
 
         # Determine which interfaces pair up into good circuits in G
         paired_interfaces = ((local_node_name, remote_node_name, data) for
@@ -807,7 +895,7 @@ class Model(MasterModel):
         """
 
         # Define a networkx DiGraph to find the path
-        G = self._make_weighted_network_graph_mdg(include_failed_circuits=include_failed_circuits, needed_bw=needed_bw)
+        G = self._make_weighted_network_graph(include_failed_circuits=include_failed_circuits, needed_bw=needed_bw)
 
         # Define the Model-style path to be built
         converted_path = dict()
@@ -836,7 +924,7 @@ class Model(MasterModel):
         """
 
         # Define a networkx DiGraph to find the path
-        G = self._make_weighted_network_graph_mdg(include_failed_circuits=False, needed_bw=needed_bw)
+        G = self._make_weighted_network_graph(include_failed_circuits=False, needed_bw=needed_bw)
 
         # Define the Model-style path to be built
         converted_path = dict()
@@ -853,6 +941,7 @@ class Model(MasterModel):
                 model_path = self._convert_nx_path_to_model_path(path)
                 converted_path['path'].append(model_path)
                 converted_path['cost'] = nx.shortest_path_length(G, source_node_name, dest_node_name, weight='cost')
+
             return converted_path
         except BaseException:
             return converted_path
@@ -1039,52 +1128,6 @@ class Model(MasterModel):
         for interface in self.interface_objects:
             pprint(interface)
             print()
-
-    # def _make_weighted_network_graph_mdg(self, include_failed_circuits=True, needed_bw=0, rsvp_required=False):
-    #     """
-    #     Returns a networkx weighted networkx multidigraph object from
-    #     the input Model object
-    #
-    #     :param include_failed_circuits: include interfaces from currently failed
-    #     circuits in the graph?
-    #     :param needed_bw: how much reservable_bandwidth is required?
-    #     :param rsvp_required: True|False; only consider rsvp_enabled interfaces?
-    #
-    #     :return: networkx multidigraph with edges that conform to the needed_bw and
-    #     rsvp_required parameters
-    #     """
-    #
-    #     G = nx.MultiDiGraph()
-    #
-    #     # Get all the edges that meet 'failed' and 'reservable_bw' criteria
-    #     if include_failed_circuits is False:
-    #         considered_interfaces = (interface for interface in self.interface_objects
-    #                                  if (interface.failed is False and
-    #                                      interface.reservable_bandwidth >= needed_bw))
-    #     elif include_failed_circuits is True:
-    #         considered_interfaces = (interface for interface in self.interface_objects
-    #                                  if interface.reservable_bandwidth >= needed_bw)
-    #
-    #     if rsvp_required is True:
-    #         edge_names = ((interface.node_object.name,
-    #                        interface.remote_node_object.name,
-    #                        {'cost': interface.cost, 'interface': interface, 'circuit_id': interface.circuit_id})
-    #                       for interface in considered_interfaces
-    #                       if interface.rsvp_enabled is True)
-    #     else:
-    #         edge_names = ((interface.node_object.name,
-    #                        interface.remote_node_object.name,
-    #                        {'cost': interface.cost, 'interface': interface, 'circuit_id': interface.circuit_id})
-    #                       for interface in considered_interfaces)
-    #
-    #     # Add edges to networkx DiGraph
-    #     G.add_edges_from(edge_names)
-    #
-    #     # Add all the nodes
-    #     node_name_iterator = (node.name for node in self.node_objects)
-    #     G.add_nodes_from(node_name_iterator)
-    #
-    #     return G
 
     def _make_weighted_network_graph_routed_lsp(self, lsp, needed_bw=0):
         """
